// backend/src/opportunities/dto/update-opportunity.dto.ts
import { OpportunityStage } from '../../generated/prisma/enums.js';

export class UpdateOpportunityDto {
  companyId?: string;
  contactId?: string;
  ownerId?: string;

  name?: string;
  description?: string;

  amount?: number;
  currency?: string;

  stage?: OpportunityStage;
  probability?: number;

  expectedCloseDate?: string;
  notes?: string;

  isArchived?: boolean;
}